`clang-tidy`: resolve `bugprone-narrowing-conversions` (#497)
* configwrap.cpp: use static_cast and add overflow checks for cvalue_size
* test/python/all_config.py: add cache-hit assertions to cover configwrap.cpp lines 62-65
* Address @knoepfel review: refactor helpers, fix real_time division, clean up casts
* Decoration required to prevent narrowing

---------

+++ test/form/toy_tracker.cpp
@@ -10,9 +10,10 @@ std::vector<TrackStart> ToyTracker::operator()()
   int32_t const npx = generateRandom() % m_maxTracks;
   std::vector<TrackStart> points(npx);
   for (int nelement = 0; nelement < npx; ++nelement) {
-    points[nelement] = TrackStart(float(generateRandom()) / random_max,
-                                  float(generateRandom()) / random_max,
-                                  float(generateRandom()) / random_max);
+    points[nelement] =
+      TrackStart(static_cast<float>(generateRandom()) / static_cast<float>(random_max),
+                 static_cast<float>(generateRandom()) / static_cast<float>(random_max),
+                 static_cast<float>(generateRandom()) / static_cast<float>(random_max));
   }
 
   return points;
